docs(graphql/quick-start): Clarify use of generate-typings script
This clarifies that the generate-typings.ts file needs to live within
the src subfolder of the project and is run from the project's root
directory.

Before, if the user created the file in the project's root folder, the
build output in the dist subfolder would be altered such that it
reflects the structure of the project root folder.
This in turn breaks the npm scripts, such as npm run start:dev, as it
expects main.js to live in dist/ instead of dist/src, where it ends up
if the user creates the script file in the project root folder.

Signed-off-by: Felix Kaechele <fkaechel@cisco.com>