No Schema without at least one query
Inform the user that it is impossible to start generating a schema without at least one query, in case they want to start with a mutation. It's actually a very plausible thing to want to start with a mutation, a login for exemple, so this is useful to know in the getting started.