Mention need for extending Cat interface
The following line will create not very user friendly compilation errors because of the incorrect Cat interface declaration:
```
constructor(@InjectModel('Cat') private readonly catModel: Model<Cat>) {}
```
The `Cats` example is used almost everywhere through the documentation, but it's not stated anywhere in the docs that you have to `extend` the interface from the mongoose `Document`. Surely, it can be easily fixed by looking at the official example, but that step is not quite straightforward, and, in my opinion, should be mentioned explicitly.