Add Nest CLI update command to the migration documentation
I noticed that there was no mention of how to actually go about updating from Nest version 6 to 7 and was struggling to find info on how to do this elsewhere. I thought it would be useful to add how to update using the Nest CLI so as people do not need to go hunting for information on this.