Add a section for overriding APP_* token providers
This adds a section on how to replace `APP_*` token providers in testing. This follows the description by @kamilmysliwiec in https://github.com/nestjs/nest/issues/4053 . I've used the solution in my project and it worked great.

Thanks to @jmcdo29 who pointed me to that issue when I was struggling with this.