docs(fastify,helmet,apollo): add warnings
apollo-server-fastify does not have full support for fastify v3 which
results in the playground not loading. Added a warning to either use
unstable apollo-server-fastify version that supports v3 or express.
Playground also uses CDN which requires a change to CSP if using helmet.
Added a warning containing required policy additions.

https://github.com/nestjs/nest/issues/6180