Update Disable Cache Expiration Docs
The documentation says to set the `ttl` to `null` should disable the cache. The TS types do not allow it to be null, and I found that actually setting it to null does not disable the cache expiration. It looks like setting the ttl to `0` is actually the way to disable the expiration, at least for the memory store and some other stores.

https://github.com/BryanDonovan/node-cache-manager/issues/107

https://github.com/BryanDonovan/node-cache-manager/issues/147