fix(content): remove JwtModule from exports
In the JWT functionality chapter in Authentication page it is needed
to remove the JwtModule from exports in the auth/auth.module example.

The JwtService is only used in the auth scope (by AuthService or
JwtStrategy) so after this commit it is removed from the exports
property in AuthModule decorator to prevent misleading interpretation
in the docs.

Fixes #1969