feat: mention the different registration methods
I've added in a segment about forRoot, forFeature, and register. I've seen
this question pop up several times on Discord, in GitHub Issues, and in
StackOverflow. I figured it was time for us to try and codify a reason
for the naming. Of course, there's no hard rule here, but the guideline
is suggested and it could become community standard which would be a big
win for the mental load of creating and using new modules.