Update interceptors.md
The definition of "interceptors" is a bit flawed, using "which" implies that the main requirement for a class to be an interceptor is that it needs to be annotated with the `@injectable` decorator. Which is not entirely true, interceptors can exist without the `@injectable` decorator but it must implement the `NestInterceptor` interface. For clarity, i have changed ",which" to an "and" as it better defines an interceptor.