fix: remove outdated section about shutdown hooks with prisma
This section has not been relevant since the introduction of the library
engine as the default in Prisma 3. Now, in Prisma 5, the code example
does not compile anymore since the `beforeExit` event was removed from
Prisma Client (except when generating the client with the binary engine
type, which most users should not do unless they have good reasons).

Note that even in Prisma 2 or when switching to the binary engine, using
`beforeExit` would've only been necessary to be able to run database
queries in lifecycle hooks, and wouldn't be necessary otherwise. The
justification for using `beforeExit` that was present in this section
was not fully accurate because while the bug it refers to
(https://github.com/prisma/prisma/issues/3773) was fixed together with
the introduction of `beforeExit` hook (which I believe led to the
confusion), those were not strictly related changes.