docs(cookies.md): update cookie-parser import syntax
the previous syntax did not work for me and just produced the following error:
```
src/main.ts:47:11 - error TS2349: This expression is not callable.
  Type 'typeof cookieParser' has no call signatures.
```
when I switched up to the new import call it worked fine. This will save time for the next person that would have wasted their time on this :)