docs(sql.md): Add documentation for testing with database
This was requested in #246 and since I was recently struggling to set this up in one of my own
projects, I have it fresh in memory how to do this. If this was in the documentation when I was
implementing this, it would have saved me a lot of time. There are many dodgy approaches to testing
NestJS applications with a test database on the Internet and an official recommended way would be
most helpful.

Resolves #246